Fixed a bug with the template conditional expressions.
If first entry was a string it would assume that all the following
entries were also strings.
It has been rewritten now so that it loops over each entry and then
determines whether it is a string, hash or array and whether
the values following is a string or array.