Release v2.18.0 - pam::service can specify array of lines and ensure
This release adds functionality to pam::service so that you can set the
ensure to present or absent, allowing you to remove entries. It also
adds the ability to specify lines as an array, so the inputs could be an
array of lines or the entire contents or nothing at all.